call registry from handlers (#389)
Calls registry from the handler functions to push info

Observations - inner source has a lot of extra log lines in the handler
functions. pkg/clusters/types has references to idps client that need to
be added in the inner source commit, and the idps client and resolvers
need to be updated in the inner source commit as well.

---------

Signed-off-by: Ryan Tay <ryan_tay@intuit.com>
Co-authored-by: Ryan Tay <ryan_tay@intuit.com>