MESH-6401: handle exportTo on in-cluster VS (#402)
### Checklist
🚨 Please review this repository's [contribution
guidelines](./CONTRIBUTING.md).

- [X] I've read and agree to the project's contribution guidelines.
- [X] I checked that my code additions will pass code linting checks and
unit tests.
- [X] I updated unit and integration tests (if applicable).
- [X] I'm ready to notify the team of this contribution.

### Description
What does this change do and why?
Handle exportTo on in cluster routing vs
Admiral will process all the clusters and all the identities, and create
VS with exportTo as "admiral-sync"
Admiral will update the exportTo to the client namespace only for the
cluster and identities where VS base in-cluster routing is enabled.

Thank you!

---------

Signed-off-by: vjoshi3 <vrushali_joshi@intuit.com>
Co-authored-by: vjoshi3 <vrushali_joshi@intuit.com>